refactor 303: 리뷰 조회 시 이미지 로딩 N+1 문제 해결

+++ backend/src/main/java/com/mysite/knitly/domain/product/review/controller/ReviewController.java
@@ -24,7 +24,7 @@ public class ReviewController {
             @PathVariable Long productId,
             @RequestParam UUID userId,
             @RequestParam("content") String content,
-            @RequestParam("rating") Byte rating,
+            @RequestParam("rating") Integer rating,
             @RequestParam(value = "images", required = false) List<MultipartFile> images
     ) {
         ReviewCreateRequest request = new ReviewCreateRequest(rating, content, images);
